requirements.txt
django
gunicorn
python-dotenv
flake8
djangorestframework
psycopg2